Cleanup pass

+++ src/Nazara/Utility/Formats/OBJParser.cpp
@@ -228,7 +228,7 @@ namespace Nz
 
 						if (p < 0)
 						{
-							p += m_positions.size() - 1;
+							p += static_cast<int>(m_positions.size() - 1);
 							if (p < 0)
 							{
 								Error("Vertex index out of range (" + String::Number(p) + " < 0");
@@ -239,7 +239,7 @@ namespace Nz
 
 						if (n < 0)
 						{
-							n += m_normals.size() - 1;
+							n += static_cast<int>(m_normals.size() - 1);
 							if (n < 0)
 							{
 								Error("Normal index out of range (" + String::Number(n) + " < 0");
@@ -250,7 +250,7 @@ namespace Nz
 
 						if (t < 0)
 						{
-							t += m_texCoords.size() - 1;
+							t += static_cast<int>(m_texCoords.size() - 1);
 							if (t < 0)
 							{
 								Error("Texture coordinates index out of range (" + String::Number(t) + " < 0");
